requirements.txt
fastapi
uvicorn
requests
python-multipart
pandas
pydantic
python-dotenv
